back/requirements.txt
# Core web framework and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data validation and settings
pydantic[email]==2.11.2
pydantic_settings==2.6.1

# Pydantic AI framework (core dependency)
pydantic-ai-slim[openai]==0.0.49

# AI and ML
openai==1.67.0
google-generativeai==0.3.2
tiktoken

# LangChain ecosystem
langchain==0.2.*
langgraph==0.0.*

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
clerk_backend_api

# HTTP and utilities
aiohttp==3.8.6
python-multipart==0.0.6
python-dotenv==1.0.0

# Image processing
pillow==10.1.0

# External services
svix

# Telegram Bot
python-telegram-bot==21.7